Create audio_responser.py
This code sends a POST request to the /command endpoint of your Flask web application running on localhost with the command parameter set to What's the weather like today?. It then checks if the request was successful by verifying the HTTP status code of the response, and prints the summarized response if the request was successful. If the request fails, it prints an error message with the HTTP status code.

You can use this code as a starting point for integrating your mobile application with the backend web application you created using Flask. You'll need to modify the url variable to match the URL of your own web application, and modify the command variable to be set by the user input from the mobile application.